+++ client/src/containers/App/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import './App.css';
-import api from './fakeApi';
+import api from '../../fakeApi';
 
 class App extends Component {
   constructor(props) {
